adds state vector methods
